Move LEAN_VERSION_RE to neutral bubble/lean.py module
Addresses Codex review feedback: the constant was in hooks/lean.py,
creating an upward dependency from builder and commands into the hook
layer. Also switches the one remaining .match() call to .fullmatch()
for consistency (no behavioral change since the regex is anchored).
